feat: completed all w2-friday
