+++ run.sh
@@ -20,11 +20,11 @@ function install {
 }
 
 function run {
-    AWS_PROFILE=cloud-course S3_BUCKET_NAME="some-bucket" uvicorn files_api.main:create_app --reload    
+    AWS_PROFILE=cloud-course S3_BUCKET_NAME="some-bucket" uvicorn files_api.main:create_app --reload
 }
 
 function run-mock {
-    
+
     # Start moto.server in the background on localhost:5000
     moto_server -p 5000 &
 
@@ -36,10 +36,10 @@ function run-mock {
     # create a bucket called "some-bucket" using the mocked aws server
     aws s3 mb "s3://$S3_BUCKET_NAME"
 
-    uvicorn files_api.main:create_app --reload  
-    
+    uvicorn files_api.main:create_app --reload
+
     # Process may not terminate, so can use this command at the terminal
-    #lsof -i :5000 | grep LISTEN | awk '{print $2}' | xargs kill -9 
+    #lsof -i :5000 | grep LISTEN | awk '{print $2}' | xargs kill -9
 }
 
 # run linting, formatting, and other static code quality tools
