+++ src/files_api/main.py
@@ -1,176 +1,17 @@
 import os
-from ast import Str
-from datetime import datetime
-from typing import (
-    List,
-    Optional,
-)
 
-from fastapi import (
-    Depends,
-    FastAPI,
-    Response,
-    UploadFile,
-    status,
-)
-from fastapi.responses import StreamingResponse
-from pydantic import BaseModel
+from fastapi import FastAPI
 
-from files_api.s3.delete_objects import delete_s3_object
-from files_api.s3.read_objects import (
-    fetch_s3_object,
-    fetch_s3_objects_metadata,
-    fetch_s3_objects_using_page_token,
-    object_exists_in_s3,
-)
-from files_api.s3.write_objects import upload_s3_object
-
-#####################
-# --- Constants --- #
-#####################
-
-# APP = FastAPI()
-
-####################################
-# --- Request/response schemas --- #
-####################################
+from files_api.routes import ROUTER
+from files_api.settings import Settings
 
 
-# create/update
-class PutFileResponse(BaseModel):
-    file_path: str
-    message: str
-
-
-# read
-class FileMetadata(BaseModel):
-    file_path: str
-    last_modified: datetime
-    size_bytes: int
-
-
-# read
-class GetFilesResponse(BaseModel):
-    files: List[FileMetadata]
-    next_page_token: Optional[str]
-
-
-# read
-class GetFilesQueryParams(BaseModel):
-    page_size: int = 10
-    directory: Optional[str] = ""
-    page_token: Optional[str] = None
-
-
-# delete
-class DeleteFileResponse(BaseModel):
-    message: str
-
-
-##################
-# --- Routes --- #
-##################
-
-
-def create_app(s3_bucket_name: str = None) -> FastAPI:
+def create_app(settings: Settings | None = None) -> FastAPI:
     """Create a FastAPI application."""
-    s3_bucket_name = s3_bucket_name or os.environ["S3_BUCKET_NAME"]
+    settings = settings or Settings()
     app = FastAPI()
-    # return app
-
-    @app.put("/files/{file_path:path}")
-    async def upload_file(file_path: str, file: UploadFile, response: Response) -> PutFileResponse:
-        """Upload a file."""
-
-        file_contents: bytes = await file.read()
-
-        object_already_exists = object_exists_in_s3(bucket_name=s3_bucket_name, object_key=file_path)
-        if object_already_exists:
-            response_message = f"Existing file updated at path: /{file_path}"
-            response.status_code = status.HTTP_200_OK
-        else:
-            response_message = f"New file uploaded at path: /{file_path}"
-            response.status_code = status.HTTP_201_CREATED
-
-        upload_s3_object(
-            bucket_name=s3_bucket_name,
-            object_key=file_path,
-            file_content=file_contents,
-            content_type=file.content_type,
-        )
-
-        return PutFileResponse(
-            file_path=file_path,
-            message=response_message,
-        )
-
-    @app.get("/files")
-    async def list_files(
-        query_params: GetFilesQueryParams = Depends(),  # noqa: B008
-    ) -> GetFilesResponse:
-        """List files with pagination."""
-
-        if query_params.page_token:
-            files, next_page_token = fetch_s3_objects_using_page_token(
-                bucket_name=s3_bucket_name,
-                continuation_token=query_params.page_token,
-                max_keys=query_params.page_size,
-            )
-        else:
-            files, next_page_token = fetch_s3_objects_metadata(
-                bucket_name=s3_bucket_name,
-                prefix=query_params.directory,
-                max_keys=query_params.page_size,
-            )
-        file_metadata_objs = [
-            FileMetadata(
-                file_path=f"{item['Key']}",
-                last_modified=item["LastModified"],
-                size_bytes=item["Size"],
-            )
-            for item in files
-        ]
-
-        return GetFilesResponse(files=file_metadata_objs, next_page_token=next_page_token if next_page_token else None)
-
-    @app.head("/files/{file_path:path}")
-    async def get_file_metadata(file_path: str, response: Response) -> Response:
-        """Retrieve file metadata.
-
-        Note: by convention, HEAD requests MUST NOT return a body in the response.
-        """
-        get_object_response = fetch_s3_object(
-            bucket_name=s3_bucket_name,
-            object_key=file_path,
-        )
-        response.headers["Content-Type"] = get_object_response["ContentType"]
-        response.headers["Content-Length"] = str(get_object_response["ContentLength"])
-        response.headers["Last-Modified"] = get_object_response["LastModified"].strftime("%a, %d %b %Y %H:%M:%S GMT")
-        response.status_code = status.HTTP_200_OK
-        return response
-
-    @app.get("/files/{file_path:path}")
-    async def get_file(
-        file_path: str,
-    ) -> StreamingResponse:
-        """Retrieve a file."""
-        get_object_response = fetch_s3_object(bucket_name=s3_bucket_name, object_key=file_path)
-        return StreamingResponse(
-            content=get_object_response["Body"],
-            media_type=get_object_response["ContentType"],
-        )
-
-    @app.delete("/files/{file_path:path}")
-    async def delete_file(
-        file_path: str,
-        response: Response,
-    ) -> Response:
-        """Delete a file.
-
-        NOTE: DELETE requests MUST NOT return a body in the response."""
-        delete_s3_object(bucket_name=s3_bucket_name, object_key=file_path)
-        response.status_code = status.HTTP_204_NO_CONTENT
-        return response
+    app.state.settings = settings
+    app.include_router(ROUTER)
 
     return app
