+++ tests/unit_tests/test__main.py
@@ -3,6 +3,7 @@ from fastapi import status
 from fastapi.testclient import TestClient
 
 from files_api.main import create_app
+from files_api.settings import Settings
 from tests.consts import TEST_BUCKET_NAME
 
 # Constants for testing
@@ -14,7 +15,8 @@ TEST_FILE_CONTENT_TYPE = "text/plain"
 # Fixture for FastAPI test client
 @pytest.fixture
 def client(mocked_aws) -> TestClient:  # pylint: disable=unused-argument
-    app = create_app(s3_bucket_name=TEST_BUCKET_NAME)
+    settings = Settings(s3_bucket_name=TEST_BUCKET_NAME)
+    app = create_app(settings=settings)
     with TestClient(app) as client:
         yield client
