+++ version.txt
@@ -1,1 +1,1 @@
-v0.0.3
+v0.0.4
